requirements.txt
fastapi>=0.104.1
httpx>=0.25.2
pydantic>=2.8.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
gunicorn>=21.0.0